move all auth to common page
