+++ src/components/signupform/BaseSignupForm.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Field, Form, ErrorMessage } from "formik";
+import { Field, Form } from "formik";
 import { TextField } from "formik-material-ui";
 import Button from "@material-ui/core/Button";
 import Link from "@material-ui/core/Link";
@@ -8,13 +8,13 @@ import RouterLink from "../RouterLink";
 import Grid from "@material-ui/core/Grid";
 import { makeStyles } from "@material-ui/core/styles";
 import { string, ref } from "yup";
+import {initialValues, validationFields} from '../formSettings'
 
 
 export const baseInitialValues = {
   firstName: "",
   lastName: "",
-  email: "",
-  password: "",
+  ...initialValues,
   verifyPassword: ""
 };
 
@@ -25,9 +25,7 @@ export const baseValidationFields = {
   lastName: string()
     .max(20, "Must be 20 characters or less")
     .required("Required"),
-  email: string()
-    .email("Invalid email addresss`")
-    .required("Required"),
+  email: validationFields.email,
   password: string()
     .min(8, "Must be 8 characters or more")
     .max(40, "Must be 40 characters or less")
@@ -74,7 +72,7 @@ const fieldAttributes = {
 export default function BaseSignupForm({ component: Component, isSubmitting }) {
   const classes = useStyles();
   return (
-    <Form className={classes.form}>
+    <Form className={classes.form} noValidate>
       <Grid container spacing={2}>
         <Grid item xs={12} sm={6}>
           <Field
@@ -95,12 +93,6 @@ export default function BaseSignupForm({ component: Component, isSubmitting }) {
             {...fieldAttributes}
           />
         </Grid>
-        <Grid item xs={12} sm={6}>
-          <ErrorMessage name="firstName" component="div" />
-        </Grid>
-        <Grid item xs={12} sm={6}>
-          <ErrorMessage name="lastName" component="div" />
-        </Grid>
         <Grid item xs={12}>
           <Field
             id="email"
@@ -111,9 +103,6 @@ export default function BaseSignupForm({ component: Component, isSubmitting }) {
             {...fieldAttributes}
             />
         </Grid>
-        <Grid item xs={12}>
-          <ErrorMessage name="email" component="div" />
-        </Grid>
         {Component && <Component />}
         <Grid item xs={12}>
           <Field
@@ -121,12 +110,9 @@ export default function BaseSignupForm({ component: Component, isSubmitting }) {
             label="Password"
             type="password"
             name="password"
-            autoComplete="email"
+            autoComplete="password"
             {...fieldAttributes}
           />
-        </Grid>
-        <Grid item xs={12}>
-          <ErrorMessage name="password" component="div" />
         </Grid>
         <Grid item xs={12}>
           <Field
@@ -134,12 +120,8 @@ export default function BaseSignupForm({ component: Component, isSubmitting }) {
             label="Verify Password"
             type="password"
             name="verifyPassword"
-            autoComplete="email"
             {...fieldAttributes}
           />
-        </Grid>
-        <Grid item xs={12}>
-          <ErrorMessage name="verifyPassword" component="div" />
         </Grid>
       </Grid>
       <Button
@@ -154,7 +136,7 @@ export default function BaseSignupForm({ component: Component, isSubmitting }) {
       </Button>
       <Grid container justify="flex-end">
         <Grid item>
-          <Link component={RouterLink} to="signin" variant="body2">
+          <Link component={RouterLink} to="/auth/signin" variant="body2">
             Already have an account? Sign in
           </Link>
         </Grid>
