+++ src/components/signupform/DonorSignup.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { object } from "yup";
-import AuthForm from "./signupform/AuthForm";
-import BaseSignupForm, {baseInitialValues, baseValidationFields} from "./signupform/BaseSignupForm";
-import DonorFields, {donorInitialValues, donorValidationFields} from "./signupform/DonorFields";
+import AuthForm from "../AuthForm";
+import BaseSignupForm, {baseInitialValues, baseValidationFields} from "./BaseSignupForm";
+import DonorFields, {donorInitialValues, donorValidationFields} from "./DonorFields";
 import { signup } from "../../actions/auth";
 
 export const initialValues = {...baseInitialValues, ...donorInitialValues};
