+++ src/components/signupform/RequesterSignup.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { object } from "yup";
-import AuthForm from "./signupform/AuthForm";
+import AuthForm from "../AuthForm";
 import BaseSignupForm, {
   baseInitialValues,
   baseValidationFields
